ml-service/requirements.txt
# Whisky visual-search ML service
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4

# Detection — YOLOv8
ultralytics==8.2.103

# OCR — PaddleOCR (위스키 라벨에 강함)
paddleocr==2.8.1
paddlepaddle==2.6.1

# Embedding — OpenCLIP
open_clip_torch==2.26.1
torch==2.3.1
torchvision==0.18.1

# Vector search — FAISS
faiss-cpu==1.8.0

# Index builder: pull reference bottles from Supabase
supabase==2.7.4
httpx==0.27.2
python-dotenv==1.0.1
